Bigger part of new backprop.
git-svn-id: http://merlin.fit.vutbr.cz/svn/STK/trunk@69 1d61b05a-ac01-0410-9a19-bfe7c875843c

+++ src/SNetLib/nnet.h
@@ -16,7 +16,7 @@ namespace SNet{
       int mVectors;               ///< Number of training frames used for training so far
       int mGood;                 ///< Number of training frames which are classified as in example
       int mDiscarded;             ///< Number of training frames discarded and not used
-      int mLearnRate;            ///< Neural net training rate
+      float mLearnRate;            ///< Neural net training rate
       bool mCrossValidation;     ///< Only compute cross-validation
       int mNCache;
       Timers *mpTimers;
